Allow multiple watch paths for single target
Sometimes you'll want to trigger a single pipeline based on multiple
paths not sharing a common root. This change allows you to specify
an array of paths to watch for a single target in addition to preserving
the old behavior of a single string.

I believe this change also unsurfaced a bug with the existing diff tests
where `buildkite-agent pipeline upload` was expected to be called in the
tests but shouldn't acutally be called based on the code in the command
hook.

Signed-off-by: Elliott Davis <elliott@excellent.io>